added circle detection
the function works. we just have to make sure that it changes parameters if the function returns None. mostly it return None if it cant detect a single circle. I think good would be to try to get atleast 3-15 circles so we can find a good approximation for our center